requirements.txt
httpx~=0.24.0
nonebot_plugin_apscheduler~=0.2.0
nonebot-adapter-onebot~=2.2.2
psutil~=5.9.5
pyyaml~=6.0
pydantic~=1.10.7
Pillow~=9.5.0